Setting up structure for versioning support
I've added modules and folders to structure the API versioning.

I've also modified the routes files to support different API versions.

https://trello.com/c/1XwDiFWt/
