+++ config/routes.rb
@@ -1,4 +1,10 @@
 Rails.application.routes.draw do
   # For details on the DSL available within this file, see http://guides.rubyonrails.org/routing.html
-  resources :products
+
+  # API Definition
+  scope module: 'api' do
+    namespace :v1 do
+      resources :products
+    end
+  end
 end
