+++ spec/controllers/api/v1/products_controller_spec.rb
@@ -1,0 +1,5 @@
+require 'rails_helper'
+
+RSpec.describe Api::V1::ProductsController, type: :controller do
+
+end
