--- spec/controllers/products_controller_spec.rb
@@ -1,5 +1,0 @@
-require 'rails_helper'
-
-RSpec.describe ProductsController, type: :controller do
-
-end
